+++ src/decorator/ComponentEachElementDecorator.ts
