+++ src/decorator/DisableableModelElementDecorator.ts
