+++ src/decorator/ValuedModelElementDecorator.ts
